Move corporate subs classes to their own package
