+++ support-models/src/main/scala/com/gu/support/redemptions/CorporateCustomer.scala
@@ -1,0 +1,5 @@
+package com.gu.support.redemptions
+
+import com.gu.support.redemptions.redemptions.{CorporateAccountId, RedemptionCode}
+
+case class CorporateCustomer(accountId: CorporateAccountId, name: String, redemptionCode: RedemptionCode)
