+++ support-models/src/main/scala/com/gu/support/workers/states/CreatePaymentMethodState.scala
@@ -6,7 +6,7 @@ import com.gu.support.promotions.PromoCode
 import com.gu.support.workers.{User, _}
 import org.joda.time.LocalDate
 import com.gu.support.encoding.CustomCodecs.{decodeLocalTime, encodeLocalTime}
-import com.gu.support.workers.redemption.RedemptionData
+import com.gu.support.redemptions.RedemptionData
 
 case class CreatePaymentMethodState(
   requestId: UUID,
