+++ support-models/src/main/scala/com/gu/support/workers/states/CreateZuoraSubscriptionState.scala
@@ -7,7 +7,7 @@ import com.gu.support.promotions.PromoCode
 import com.gu.support.workers.{PaymentMethod, SalesforceContactRecord, User, _}
 import org.joda.time.LocalDate
 import com.gu.support.encoding.CustomCodecs.{decodeLocalTime, encodeLocalTime}
-import com.gu.support.workers.redemption.RedemptionData
+import com.gu.support.redemptions.RedemptionData
 
 case class CreateZuoraSubscriptionState(
   requestId: UUID,
