+++ support-models/src/main/scala/com/gu/support/workers/states/FailureHandlerState.scala
@@ -3,7 +3,7 @@ package com.gu.support.workers.states
 import java.util.UUID
 
 import com.gu.support.promotions.PromoCode
-import com.gu.support.workers.redemption.RedemptionData
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers.{User, _}
 import org.joda.time.LocalDate
