+++ support-models/src/main/scala/com/gu/support/workers/states/SendAcquisitionEventState.scala
@@ -4,7 +4,7 @@ import java.util.UUID
 
 import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.promotions.PromoCode
-import com.gu.support.workers.redemption.RedemptionData
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers.{PaymentMethod, User, _}
 import org.joda.time.LocalDate
