+++ support-models/src/main/scala/com/gu/support/zuora/api/SubscriptionData.scala
@@ -7,7 +7,7 @@ import com.gu.support.encoding.Codec._
 import com.gu.support.encoding.CustomCodecs.{decodeDateTime, encodeDateTime, monthDecoder, _}
 import com.gu.support.encoding.JsonHelpers._
 import com.gu.support.promotions.PromoCode
-import com.gu.support.workers.redemption.{CorporateAccountId, RedemptionCode}
+import com.gu.support.redemptions.redemptions.{CorporateAccountId, RedemptionCode}
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder, Json}
 import org.joda.time.{LocalDate, Months}
