+++ support-workers/src/main/scala/com/gu/emailservices/ContributionEmailFields.scala
@@ -3,8 +3,8 @@ package com.gu.emailservices
 import com.gu.emailservices.SubscriptionEmailFieldHelpers.{formatDate, hyphenate, mask}
 import com.gu.i18n.Currency
 import com.gu.salesforce.Salesforce.SfContactId
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers._
-import com.gu.support.workers.redemption.RedemptionData
 import org.joda.time.DateTime
 
 case class ContributionEmailFields(
