+++ support-workers/src/main/scala/com/gu/emailservices/DigitalPackEmailFields.scala
@@ -4,8 +4,8 @@ import com.gu.emailservices.SubscriptionEmailFieldHelpers._
 import com.gu.i18n.Currency
 import com.gu.salesforce.Salesforce.SfContactId
 import com.gu.support.promotions.Promotion
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers._
-import com.gu.support.workers.redemption.RedemptionData
 
 // Output Json should look like this:
 //
