+++ support-workers/src/main/scala/com/gu/emailservices/GuardianWeeklyEmailFields.scala
@@ -5,8 +5,8 @@ import com.gu.i18n.Currency
 import com.gu.salesforce.Salesforce.SfContactId
 import com.gu.support.catalog.FulfilmentOptions
 import com.gu.support.promotions.Promotion
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers._
-import com.gu.support.workers.redemption.RedemptionData
 import org.joda.time.LocalDate
 
 import scala.collection.immutable
