+++ support-workers/src/main/scala/com/gu/emailservices/PaperFieldsGenerator.scala
@@ -3,8 +3,8 @@ package com.gu.emailservices
 import com.gu.i18n.Currency
 import com.gu.salesforce.Salesforce.SfContactId
 import com.gu.support.promotions.Promotion
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers._
-import com.gu.support.workers.redemption.RedemptionData
 import org.joda.time.LocalDate
 
 //noinspection ScalaStyle
