+++ support-workers/src/main/scala/com/gu/monitoring/LambdaExecutionResult.scala
@@ -6,9 +6,9 @@ import com.gu.i18n.Country
 import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.encoding.JsonHelpers._
 import com.gu.support.promotions.PromoCode
+import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers.CheckoutFailureReasons.CheckoutFailureReason
 import com.gu.support.workers._
-import com.gu.support.workers.redemption.RedemptionData
 import io.circe.generic.semiauto.deriveEncoder
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder}
